reddit/data-pipeline/requirements.txt
praw==7.7.1
pandas>=2.1.4
scikit-learn>=1.4.0
numpy>=1.26.0
nltk>=3.8.1
python-dotenv>=1.0.0
networkx>=3.2.1
openai>=1.0.0